requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
openS3-sdk==0.1.0
requests==2.32.3
urllib3==2.4.0
flask==2.3.3
werkzeug==2.3.7
flask-wtf==1.1.1
wtforms==3.0.1
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
